refactor(widget): convert set_vbox_layout/set_hbox_layout to use spacing enum
Updates the base widget class convenience methods to accept spacing enum
instead of integers, completing the spacing API refactoring.

Changes to widget.hh:
- set_vbox_layout() now accepts spacing enum parameter (default: spacing::none)
- set_hbox_layout() now accepts spacing enum parameter (default: spacing::none)
- Added static resolve_spacing() helper method for theme-based resolution
- Updated documentation to reflect semantic spacing usage

Test updates (15 files):
- Replaced all set_vbox_layout(int) calls with spacing enum equivalents
- Replaced all set_hbox_layout(int) calls with spacing enum equivalents
- Updated assertions to match theme-resolved values:
  * spacing::tiny resolves to 0 in test theme
  * spacing::small resolves to 1 in test theme
  * spacing::medium resolves to 1 in test theme
  * spacing::large resolves to 2 in test theme
  * spacing::xlarge resolves to 3 in test theme
- Fixed test expectations to use correct spacing values

Documentation updates:
- Updated panel.hh examples to use spacing::medium instead of hardcoded 5

All 1554 tests pass with 8858 assertions.

Breaking change: set_vbox_layout() and set_hbox_layout() no longer accept
integer spacing values. Use spacing enum instead.

+++ unittest/widgets/test_scroll_view.cc
@@ -448,7 +448,7 @@ TEST_CASE_FIXTURE(ui_context_fixture<test_canvas_backend>, "scroll_view - Nested
     // Group box (like scrolling_section in the layout tab)
     auto* group = content->template emplace_child<group_box>();
     group->set_title("Scrolling Section");
-    group->set_vbox_layout(static_cast<int>(spacing::tiny));
+    group->set_vbox_layout(spacing::tiny);
 
     // Add a label inside group_box
     group->template emplace_child<label>("Label inside group_box");
